Added Unit Test project
