+++ HotRS.Tools.Core.Tests/AssemblyExtensionsTests.cs
@@ -1,0 +1,32 @@
+﻿namespace HotRS.Tools.Core.Tests;
+
+[ExcludeFromCodeCoverage]
+
+public static class AssemblyExtensionsTests
+{
+    [Test]
+    public static void GetTextFileFromAssemblySuccess()
+    {
+        var fileName = "TestData.TestData.json";
+        var actual = Assembly.GetExecutingAssembly().GetTextFileFromAssembly(fileName);
+        Assert.That(actual.Length, Is.GreaterThan(0));
+
+        var convertedActual = JsonConvert.DeserializeObject<TestData.TestData>(actual);
+        Assert.That(convertedActual.BoolValue, Is.EqualTo(true));
+        Assert.That(convertedActual.EnumValue, Is.EqualTo(TestData.TestEnum.EnumEntry3));
+        Assert.That(convertedActual.IntValue, Is.GreaterThan(-1));
+        Assert.That(convertedActual.ListValue.Count, Is.GreaterThan(1));
+        Assert.That(convertedActual.StringValue, Is.Not.Empty);
+    }
+
+    [Test]
+    public static void GetTextFileFromAssemblyNotFound()
+    {
+        var fileName = "invalidfilename";
+
+        Assert.That(() => Assembly.GetExecutingAssembly().GetTextFileFromAssembly(fileName),
+            Throws.Exception.TypeOf<FileNotFoundException>()
+            .With.Property("Message").
+            EqualTo($"{fileName} does not exist in the assembly."));
+    }
+}
