+++ HotRS.Tools.Core.Tests/CollectionExtensionsTests.cs
@@ -1,0 +1,37 @@
+﻿namespace HotRS.Tools.Core.Tests;
+
+[ExcludeFromCodeCoverage]
+
+public static class CollectionExtensionsTests
+{
+	[Test]
+	public static void IsNullOrEmptyIsEmpty()
+	{
+		var expected = BuildList(0);
+		Assert.That(expected.IsNullOrEmpty(), Is.True);
+	}
+
+	[Test]
+	public static void IsNullOrEmptyIsNull()
+	{
+		List<string> expected = null;
+		Assert.That(expected.IsNullOrEmpty(), Is.True);
+	}
+
+	[Test]
+	public static void IsNullOrEmptyIsFalse()
+	{
+		var expected = BuildList(2);
+		Assert.That(expected.IsNullOrEmpty(), Is.False);
+	}
+
+	private static List<string> BuildList(int count)
+	{
+		var result = new List<string>();
+		for(var ndx = 0; ndx < count; ndx++)
+		{
+			result.Add($"Item {ndx}");
+		}
+		return result;
+	}
+}
