+++ HotRS.Tools.Core.Tests/Extensions/AppSettings.cs
@@ -1,0 +1,9 @@
+﻿namespace HotRS.Tools.Core.Tests.Extensions;
+
+[ExcludeFromCodeCoverage]
+public class AppSettings
+{
+    public List<string> Array1 { get; set; }
+    public string Setting1 { get; set; }
+    public string Setting2 { get; set; }
+}
