+++ HotRS.Tools.Core.Tests/Extensions/ConfigurationExtensionsTests.cs
@@ -1,0 +1,57 @@
+﻿namespace HotRS.Tools.Core.Extensions.Tests;
+
+[ExcludeFromCodeCoverage]
+[TestFixture()]
+public class ConfigurationExtensionsTests
+{
+    [Test()]
+    public void CleanUpJSONConfigsTest()
+    {
+        //starting with the default configuration, add another copy of the Env1 file.
+        var config = BuildConfigWithDup();
+        Assert.That(config.Providers.Count() == 4);
+        config.CleanUpJSONConfigs();
+        Assert.That(config.Providers.Count() == 3);
+
+    }
+
+    [Test()]
+    public void CleanUpJSONConfigsThrowsExceptionForChainedConfigurationProviders()
+    {
+        //starting with the default configuration, add another copy of the Env1 file.
+        var config = new ConfigurationBuilder()
+            .AddConfiguration(BuildConfig())
+            .AddJsonFile(@"extensions\appsettings.Env1.json")
+            .Build();
+        Assert.That(() => config.CleanUpJSONConfigs(), Throws.TypeOf<NotImplementedException>()
+            .With.Message.EqualTo("The CleanupJsonConfig extension cannot handle Chained configurations."));
+    }
+
+    [Test]
+    public void DumpTest()
+    {
+        //starting with the default configuration, add another copy of the Env1 file.
+        var config = new ConfigurationBuilder()
+            .AddConfiguration(BuildConfig())
+            .AddJsonFile(@"extensions\appsettings.Env1.json")
+            .Build();
+        var actual = config.Dump();
+        Assert.That(actual.Count > 0);
+        Console.WriteLine(JsonConvert.SerializeObject(actual, Formatting.Indented));
+    }
+
+    private static IConfigurationRoot BuildConfig() =>
+        new ConfigurationBuilder()
+            .AddJsonFile(@"extensions\appsettings.json")
+            .AddJsonFile(@"extensions\appsettings.Env1.json")
+            .AddJsonFile(@"extensions\appsettings.Env2.json")
+            .Build();
+
+    private static IConfigurationRoot BuildConfigWithDup() =>
+        new ConfigurationBuilder()
+            .AddJsonFile(@"extensions\appsettings.json")
+            .AddJsonFile(@"extensions\appsettings.Env1.json")
+            .AddJsonFile(@"extensions\appsettings.Env2.json")
+            .AddJsonFile(@"extensions\appsettings.Env1.json")
+            .Build();
+}
