+++ HotRS.Tools.Core.Tests/Extensions/appsettings.Env1.json
