+++ HotRS.Tools.Core.Tests/Extensions/appsettings.Env2.json
