+++ HotRS.Tools.Core.Tests/Extensions/appsettings.json
