+++ HotRS.Tools.Core.Tests/GlobalUsings.cs
@@ -1,0 +1,23 @@
+﻿global using HotRS.Tools.Core.Extensions;
+global using Microsoft.Extensions.Configuration;
+global using Microsoft.OpenApi.Models;
+global using Newtonsoft.Json;
+global using NUnit.Framework;
+global using Swashbuckle.AspNetCore.SwaggerGen;
+global using System;
+global using System.Collections.Generic;
+global using System.Diagnostics.CodeAnalysis;
+global using System.Linq;
+global using System.Net;
+global using static HotRS.Tools.Core.Extensions.EnumExtensions;
+global using System.IO;
+global using System.Reflection;
+global using HotRS.Tools.Core.Utilities.JSON;
+global using HotRS.Tools.Core.Exceptions;
+global using HotRS.Tools.Core.Helpers.Office;
+global using System.Diagnostics;
+global using System.Globalization;
+global using System.ComponentModel.DataAnnotations;
+global using HotRS.Tools.Core.Helpers.Misc;
+global using HotRS.Tools.Core.Tests.TestData;
+
