--- RK.HotRS.ToolsCore.Tests/Extensions/AppSettings.cs
@@ -1,13 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
-
-namespace RK.HotRS.ToolsCore.Tests.Extensions
-{
-    [ExcludeFromCodeCoverage]
-    public class AppSettings
-    {
-        public List<string> Array1 { get; set; }
-        public string Setting1 { get; set; }
-        public string Setting2 { get; set; }
-    }
-}
