--- RK.HotRS.ToolsCore.Tests/Middleware/SwaggerTools/FormFileParameterTests.cs
@@ -1,50 +1,0 @@
-﻿using Microsoft.OpenApi.Models;
-using NUnit.Framework;
-using RK.HotRS.ToolsCore.Middleware.SwaggerTools;
-using Swashbuckle.AspNetCore.SwaggerGen;
-using System;
-using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
-using System.Linq;
-using System.Text;
-
-namespace RK.HotRS.ToolsCore.Middleware.SwaggerTools.Tests
-{
-    [TestFixture()]
-    [ExcludeFromCodeCoverage]
-    public class FormFileParameterTests
-    {
-        [Test()]
-        public void ApplyTestNullOperationFails()
-        {
-            var openApiOperation = new OpenApiOperation();
-            var operationFilterContext = new OperationFilterContext(null, null, null, null);
-            var obj2Test = new FormFileParameter();
-            Assert.That(() => obj2Test.Apply(null, operationFilterContext),
-                Throws.Exception.TypeOf<ArgumentNullException>()
-                .With.Message.EqualTo("Value cannot be null. (Parameter 'operation')"));
-        }
-
-        [Test()]
-        public void ApplyTestNullOperationNotMyOperation()
-        {
-            var openApiOperation = new OpenApiOperation { OperationId = "YourOperation"};
-            var operationFilterContext = new OperationFilterContext(null, null, null, null);
-            var obj2Test = new FormFileParameter();
-            Assert.DoesNotThrow(() => obj2Test.Apply(openApiOperation, operationFilterContext));
-        }
-
-        [Test()]
-        public void ApplyTestNullOperationMyOperation()
-        {
-            var openApiOperation = new OpenApiOperation { OperationId = "MyOperation" };
-            var operationFilterContext = new OperationFilterContext(null, null, null, null);
-            var obj2Test = new FormFileParameter();
-            Assert.DoesNotThrow(() => obj2Test.Apply(openApiOperation, operationFilterContext));
-            Assert.That(openApiOperation.Parameters.Count > 0);
-            var firstParam = openApiOperation.Parameters.First();
-            Assert.That(firstParam.Name.Equals("formfile", StringComparison.OrdinalIgnoreCase));
-            Assert.That(firstParam.Schema.Type.Equals("file", StringComparison.OrdinalIgnoreCase));
-        }
-    }
-}
